+++ kernel/di.go
@@ -6,24 +6,9 @@ import (
 	"log"
 )
 
-// Container will provide access to the global Service Container
-var Container *dig.Container
-
-// BuildCustomContainer provides a service container with custom services
+// BuildCustomContainer provides a service container with custom services.
+// It returns a container that will only be user on the HTTP controllers.
 func BuildCustomContainer() *dig.Container {
-	container := dig.New()
-
-	for _, s := range CustomServices.List {
-		if err := container.Provide(s); err != nil {
-			log.Fatal(err)
-		}
-	}
-
-	return container
-}
-
-// BuildSystemContainer provide the global service container
-func BuildSystemContainer() *dig.Container {
 	container := dig.New()
 
 	for _, s := range Services.List {
@@ -32,8 +17,20 @@ func BuildSystemContainer() *dig.Container {
 		}
 	}
 
+	return container
+}
+
+// BuildSingletonContainer provide the global service container
+func BuildSingletonContainer() *dig.Container {
+	container := dig.New()
+
+	for _, s := range SingletonServices.List {
+		if err := container.Provide(s); err != nil {
+			log.Fatal(err)
+		}
+	}
+
 	injectBasicEntities(container)
-	Container = container
 
 	return container
 }