+++ kernel/register.go
@@ -18,15 +18,13 @@ var (
 			// Here is where you've to register your custom controller
 		},
 	}
-	Services = register.ServiceRegister{
+	SingletonServices = register.ServiceRegister{
 		List: []interface{}{
 			RetrieveAppConf,
 			CreateSessionStore,
-			GetHttpServer,
-			WebRouter,
 		},
 	}
-	CustomServices = register.ServiceRegister{
+	Services = register.ServiceRegister{
 		List: []interface{}{},
 	}
 	Models = register.ModelRegister{